requirements.txt
attrs==20.3.0
iniconfig==1.1.1
numpy==1.19.4
packaging==20.8
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.1
toml==0.10.2
